fix(Produto): Atualizando atributos da branch /marcos

+++ src/main/java/ammm_tech_labs/e_commerce_mvp/dto/list/ListarProdutoDTO.java
@@ -1,5 +1,6 @@
 package ammm_tech_labs.e_commerce_mvp.dto.list;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 import ammm_tech_labs.e_commerce_mvp.models.Produto;
@@ -7,7 +8,7 @@ import ammm_tech_labs.e_commerce_mvp.models.Produto;
 public record ListarProdutoDTO(
     UUID id,
     String nome,
-    Double preco
+    BigDecimal preco
 ) {
     public ListarProdutoDTO(Produto produto) {
         this(
